+++ Xqueue/Maileon/Model/Api/TestSendAbandonedCartsEmails.php
@@ -1,5 +1,5 @@
 <?php
- 
+
 namespace Xqueue\Maileon\Model\Api;
 
 use Magento\Framework\App\Config\ScopeConfigInterface;
@@ -7,7 +7,7 @@ use Magento\Store\Model\ScopeInterface;
 use Xqueue\Maileon\Model\Maileon\ContactCreate;
 use Xqueue\Maileon\Model\Maileon\TransactionCreate;
 use de\xqueue\maileon\api\client\contacts\Contact;
- 
+
 class TestSendAbandonedCartsEmails
 {
     /**
@@ -39,7 +39,7 @@ class TestSendAbandonedCartsEmails
      * @var \Magento\Framework\Message\ManagerInterface
      */
     protected $messageManager;
- 
+
     public function __construct(
         \Psr\Log\LoggerInterface $logger,
         \Magento\Framework\ObjectManagerInterface $objectManager,
@@ -59,7 +59,7 @@ class TestSendAbandonedCartsEmails
     /**
      * @inheritdoc
      */
- 
+
     public function testSendAbandonedCartsEmails($token)
     {
         $isEnabledWebhookTest = $this->isEnabledWebhookTest($token);
@@ -323,8 +323,8 @@ class TestSendAbandonedCartsEmails
         $cartItems = $this->createCartItems($quote);
 
         $content['cart.items']       = $cartItems['items'];
-        $content['cart.product_ids'] = join(',', $cartItems['productIds']);
-        $content['cart.categories']  = $cartItems['categories'];
+        $content['cart.product_ids'] = $this->sanitizeProductIdList($cartItems['productIds']);
+        $content['cart.categories']  = $this->sanitizeCategoriesList($cartItems['categories']);
         $content['cart.total']       = (float) $this->formatPrice($quote->getGrandTotal());
         $content['cart.total_tax']   = (float) $this->formatPrice($quote->getGrandTotal() - $quote->getSubtotal());
         $content['cart.currency']    = $quote->getBaseCurrencyCode();
@@ -490,7 +490,7 @@ class TestSendAbandonedCartsEmails
 
         if (!empty($categoryIds)) {
             $categories = [];
-            
+
             foreach ($categoryIds as $categoryId) {
                 $category = $this->objectManager->create('Magento\Catalog\Model\Category')->load($categoryId);
                 $categories[] = $category->getName();
@@ -523,7 +523,7 @@ class TestSendAbandonedCartsEmails
         } catch (\Exception $e) {
             $imageUrl = '';
         }
-        
+
         return $imageUrl;
     }
 
@@ -544,7 +544,7 @@ class TestSendAbandonedCartsEmails
         } catch (\Exception $e) {
             $imageUrl = '';
         }
-        
+
         return $imageUrl;
     }
 
@@ -554,7 +554,7 @@ class TestSendAbandonedCartsEmails
      * @param mixed $price
      * @return string
      */
-    protected function formatPrice(mixed $price): string
+    protected function formatPrice($price): string
     {
         return number_format(
             doubleval($price),
@@ -562,5 +562,53 @@ class TestSendAbandonedCartsEmails
             '.',
             ''
         );
+    }
+
+    /**
+     * @param string $categoriesList
+     * @return string
+     */
+    protected function sanitizeCategoriesList(string $categoriesList): string
+    {
+        if (empty($categoriesList)) {
+            return '';
+        }
+
+        $categories = explode(',', $categoriesList);
+        $categories = array_filter(array_map('trim', $categories));
+        $uniqueCategories = array_unique($categories);
+        $categoriesList = implode(',', $uniqueCategories);
+
+        return $this->sanitizeTransactionStringValue($categoriesList);
+    }
+
+    /**
+     * @param array $productIds
+     * @return string
+     */
+    protected function sanitizeProductIdList(array $productIds): string
+    {
+        if (empty($productIds)) {
+            return '';
+        }
+
+        $productIds = array_filter(array_map('trim', $productIds));
+        $uniqueProductIds = array_unique($productIds);
+        $productIdList = implode(',', $uniqueProductIds);
+
+        return $this->sanitizeTransactionStringValue($productIdList);
+    }
+
+    /**
+     * @param $value
+     * @return string
+     */
+    protected function sanitizeTransactionStringValue($value): string
+    {
+        if (!empty($value)) {
+            return mb_substr($value, 0, 1000);
+        } else {
+            return '';
+        }
     }
 }